refactor(utils): move removeDesignMdAttachmentLinks to @sokosumi/utils
